chore: WIP library refactor/reorganization, split wallet.rs out
